requirements.txt
fastapi>=0.115.12
uvicorn>=0.23.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.4.0
python-multipart>=0.0.19
pyjwt>=2.8.0
